requirements.txt
click==8.0.3
cx-Freeze==6.8.3
importlib-metadata==4.8.2
PyYAML==6.0
zipp==3.6.0
